add back button & action to history
this took some time to figure out how to do correctly when having history
embedded in a navigation controller. I wanted to add the button in
storyboard, not in code. Turns out, the way you do this is to drag
out a bar button item to the embedded controller's navigation bar,
which deceptively in ios 9 looks just like a blank space.